requirements.txt
requests==2.21.0
instabot==0.35.0
python-dotenv==0.10.1